Valid move and checkwin
Able to use a touch event to place an x or o on the board if the move is valid and able to print a message stating which player x or o won. Just need to add a canvas draw text object to the app to notify which player won. Also need to experiment with multiple views so that when the game ends the user is taken back to a screen where they can start a new game or quit the application.